chore: move files to lib folder
