+++ lib/namespaces.ts
@@ -1,6 +1,5 @@
 import namespace from "@rdfjs/namespace";
-
-const BLANK_NODE_NAMESPACE = process.env.BLANK_NODE_NAMESPACE || "http://mu.semte.ch/blank#";
+import { BLANK_NODE_NAMESPACE } from "../cfg";
 
 export const PROV = namespace("http://www.w3.org/ns/prov#");
 export const DCTERMS = namespace("http://purl.org/dc/terms/");
