+++ lib/sparql-queries.ts
@@ -1,7 +1,5 @@
 import * as RDF from "@rdfjs/types";
-import { TreeProperties, extractEndpointHeadersFromEnv, toString } from "./utils";
-import { DataFactory } from "n3";
-import { EXT } from "./namespaces";
+import { extractEndpointHeadersFromEnv, toString } from "./utils";
 import {
   MU_APPLICATION_GRAPH,
   SPARQL_AUTH_USER,
@@ -9,11 +7,9 @@ import {
   SPARQL_ENDPOINT_HEADER_PREFIX,
   SPARQL_BATCH_SIZE,
   ENABLE_SPARQL_BATCHING
-} from "./config";
-import { State } from "@treecg/actor-init-ldes-client";
+} from "../cfg";
 // @ts-ignore
 import { querySudo, updateSudo, ConnectionOptions } from "@lblod/mu-auth-sudo";
-const { quad, namedNode, variable, literal } = DataFactory;
 
 const SPARQL_ENDPOINT_HEADERS = extractEndpointHeadersFromEnv(SPARQL_ENDPOINT_HEADER_PREFIX);
 
@@ -122,85 +118,10 @@ export async function executeDeleteQuery (quads: RDF.Quad[]) {
   }
 }
 
-export async function getLatestTimestamp (baseResource: RDF.NamedNode, treeProperties: TreeProperties) {
-  const quads = [
-    quad(variable("version"), treeProperties.versionOfPath, baseResource),
-    quad(variable("version"), treeProperties.timestampPath, variable("timestamp"))
-  ];
-  const variables = [variable("timestamp")];
-  const sparqlQuery = constructSelectQuery(variables, quads, 'ORDER BY DESC(?timestamp)');
-  try {
-    const response = await query(sparqlQuery);
-    const timestamp = extractVariableFromResponse(response, "timestamp")?.shift();
-    if (timestamp) {
-      return timestamp;
-    }
-    return;
-  } catch (e) {
-    console.error(e);
-  }
-}
-
 export async function executeDeleteInsertQuery (
   quadsToDelete: RDF.Quad[],
   quadsToInsert: RDF.Quad[]
 ) {
   await executeDeleteQuery(quadsToDelete);
   await executeInsertQuery(quadsToInsert);
-}
-
-export async function fetchState (stream: RDF.NamedNode): Promise<State | undefined> {
-  const quads = [
-    quad(stream, EXT("state"), variable("state"))
-  ];
-  const variables = [variable("state")];
-  const sparqlQuery = constructSelectQuery(variables, quads);
-  try {
-    const response = await query(sparqlQuery);
-    const stateString = extractVariableFromResponse(response, "state")?.shift();
-    if (stateString) {
-      return JSON.parse(stateString);
-    }
-    return;
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-function extractVariableFromResponse (
-  response: any,
-  variable: string
-): string[] | undefined {
-  const bindings = response.results.bindings;
-  if (bindings && bindings.length) {
-    return bindings.map((binding: any) => binding[variable].value);
-  }
-}
-
-export async function getVersion (resource: RDF.NamedNode, treeProperties: TreeProperties) {
-  const quads = [quad(variable("v"), treeProperties.versionOfPath, resource)];
-  const variables = [variable("v")];
-  const sparqlQuery = constructSelectQuery(variables, quads);
-
-  try {
-    const response = await query(sparqlQuery);
-    const versionUris = extractVariableFromResponse(response, "v");
-    if (versionUris) {
-      return namedNode(versionUris[0]);
-    }
-    return;
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-export async function updateState (stream: RDF.NamedNode, state: State) {
-  const stateString = JSON.stringify(state);
-  const genericStateQuads = [
-    quad(stream, EXT("state"), variable("state"))
-  ];
-  const newStateQuads = [
-    quad(stream, EXT("state"), literal(stateString))
-  ];
-  await executeDeleteInsertQuery(genericStateQuads, newStateQuads);
 }
